tabla1

+++ app/db/migrate/20190311195853_create_empleados.rb
@@ -3,14 +3,14 @@ class CreateEmpleados < ActiveRecord::Migration[5.0]
     create_table :empleados do |t|
         t.text :nombres
         t.text :apellidos
-        t.int8 :cedula
-        t.int8 :celular
+        t.integer :cedula
+        t.integer :celular
         t.timestamps
     end
     create_table :roles do |t|
         t.belongs_to :supplier, index: { unique: true }, foreign_key: true
         t.text :descripcion
-        t.int8 :sueldo
+        t.integer :sueldo
         t.timestamps
     end       
   end
